+++ src/tests/integration/server.itest.js
@@ -45,281 +45,6 @@ describe('/api', () => {
   });
 });
 
-describe('/api/features', () => {
-  it('should list features', async () => {
-    const res = await request(app).get('/api/features');
-    expect(res.statusCode).toEqual(200);
-    expect(res.body.results.length).not.toEqual(0);
-  });
-  describe('with name query', () => {
-    it('returns the named object', async () => {
-      const indexRes = await request(app).get('/api/features');
-      const name = indexRes.body.results[1].name;
-      const res = await request(app).get(`/api/features?name=${name}`);
-      expect(res.statusCode).toEqual(200);
-      expect(res.body.results[0].name).toEqual(name);
-    });
-
-    it('is case insensitive', async () => {
-      const indexRes = await request(app).get('/api/features');
-      const name = indexRes.body.results[1].name;
-      const queryName = name.toLowerCase();
-      const res = await request(app).get(`/api/features?name=${queryName}`);
-      expect(res.statusCode).toEqual(200);
-      expect(res.body.results[0].name).toEqual(name);
-    });
-  });
-  describe('/api/features/:index', () => {
-    it('should return one object', async () => {
-      const indexRes = await request(app).get('/api/features');
-      const index = indexRes.body.results[0].index;
-      const showRes = await request(app).get(`/api/features/${index}`);
-      expect(showRes.statusCode).toEqual(200);
-      expect(showRes.body.index).toEqual(index);
-    });
-    describe('with an invalid index', () => {
-      it('should return one object', async () => {
-        const invalidIndex = 'invalid-index';
-        const showRes = await request(app).get(`/api/features/${invalidIndex}`);
-        expect(showRes.statusCode).toEqual(404);
-      });
-    });
-  });
-});
-
-describe('/api/languages', () => {
-  it('should list languages', async () => {
-    const res = await request(app).get('/api/languages');
-    expect(res.statusCode).toEqual(200);
-    expect(res.body.results.length).not.toEqual(0);
-  });
-  describe('with name query', () => {
-    it('returns the named object', async () => {
-      const indexRes = await request(app).get('/api/languages');
-      const name = indexRes.body.results[1].name;
-      const res = await request(app).get(`/api/languages?name=${name}`);
-      expect(res.statusCode).toEqual(200);
-      expect(res.body.results[0].name).toEqual(name);
-    });
-
-    it('is case insensitive', async () => {
-      const indexRes = await request(app).get('/api/languages');
-      const name = indexRes.body.results[1].name;
-      const queryName = name.toLowerCase();
-      const res = await request(app).get(`/api/languages?name=${queryName}`);
-      expect(res.statusCode).toEqual(200);
-      expect(res.body.results[0].name).toEqual(name);
-    });
-  });
-  describe('/api/languages/:index', () => {
-    it('should return one object', async () => {
-      const indexRes = await request(app).get('/api/languages');
-      const index = indexRes.body.results[0].index;
-      const showRes = await request(app).get(`/api/languages/${index}`);
-      expect(showRes.statusCode).toEqual(200);
-      expect(showRes.body.index).toEqual(index);
-    });
-    describe('with an invalid index', () => {
-      it('should return one object', async () => {
-        const invalidIndex = 'invalid-index';
-        const showRes = await request(app).get(`/api/languages/${invalidIndex}`);
-        expect(showRes.statusCode).toEqual(404);
-      });
-    });
-  });
-});
-
-describe('/api/magic-items', () => {
-  it('should list magic items', async () => {
-    const res = await request(app).get('/api/magic-items');
-    expect(res.statusCode).toEqual(200);
-    expect(res.body.results.length).not.toEqual(0);
-  });
-
-  it('should hit the cache', async () => {
-    redisClient.del('/api/magic-items');
-    const clientSet = jest.spyOn(redisClient, 'set');
-    let res = await request(app).get('/api/magic-items');
-    res = await request(app).get('/api/magic-items');
-    expect(res.statusCode).toEqual(200);
-    expect(res.body.results.length).not.toEqual(0);
-    expect(clientSet).toHaveBeenCalledTimes(1);
-  });
-
-  describe('with name query', () => {
-    it('returns the named object', async () => {
-      const indexRes = await request(app).get('/api/magic-items');
-      const name = indexRes.body.results[2].name;
-      const res = await request(app).get(`/api/magic-items?name=${name}`);
-      expect(res.statusCode).toEqual(200);
-      expect(res.body.results[0].name).toEqual(name);
-    });
-
-    it('is case insensitive', async () => {
-      const indexRes = await request(app).get('/api/magic-items');
-      const name = indexRes.body.results[2].name;
-      const queryName = name.toLowerCase();
-      const res = await request(app).get(`/api/magic-items?name=${queryName}`);
-      expect(res.statusCode).toEqual(200);
-      expect(res.body.results[0].name).toEqual(name);
-    });
-  });
-  describe('/api/magic-items/:index', () => {
-    it('should return one object', async () => {
-      const indexRes = await request(app).get('/api/magic-items');
-      const index = indexRes.body.results[0].index;
-      const showRes = await request(app).get(`/api/magic-items/${index}`);
-      expect(showRes.statusCode).toEqual(200);
-      expect(showRes.body.index).toEqual(index);
-    });
-    describe('with an invalid index', () => {
-      it('should return one object', async () => {
-        const invalidIndex = 'invalid-index';
-        const showRes = await request(app).get(`/api/magic-items/${invalidIndex}`);
-        expect(showRes.statusCode).toEqual(404);
-      });
-    });
-  });
-});
-
-describe('/api/magic-schools', () => {
-  it('should list magic items', async () => {
-    const res = await request(app).get('/api/magic-schools');
-    expect(res.statusCode).toEqual(200);
-    expect(res.body.results.length).not.toEqual(0);
-  });
-  describe('with name query', () => {
-    it('returns the named object', async () => {
-      const indexRes = await request(app).get('/api/magic-schools');
-      const name = indexRes.body.results[1].name;
-      const res = await request(app).get(`/api/magic-schools?name=${name}`);
-      expect(res.statusCode).toEqual(200);
-      expect(res.body.results[0].name).toEqual(name);
-    });
-
-    it('is case insensitive', async () => {
-      const indexRes = await request(app).get('/api/magic-schools');
-      const name = indexRes.body.results[1].name;
-      const queryName = name.toLowerCase();
-      const res = await request(app).get(`/api/magic-schools?name=${queryName}`);
-      expect(res.statusCode).toEqual(200);
-      expect(res.body.results[0].name).toEqual(name);
-    });
-  });
-  describe('/api/magic-schools/:index', () => {
-    it('should return one object', async () => {
-      const indexRes = await request(app).get('/api/magic-schools');
-      const index = indexRes.body.results[0].index;
-      const showRes = await request(app).get(`/api/magic-schools/${index}`);
-      expect(showRes.statusCode).toEqual(200);
-      expect(showRes.body.index).toEqual(index);
-    });
-    describe('with an invalid index', () => {
-      it('should return one object', async () => {
-        const invalidIndex = 'invalid-index';
-        const showRes = await request(app).get(`/api/magic-schools/${invalidIndex}`);
-        expect(showRes.statusCode).toEqual(404);
-      });
-    });
-  });
-});
-
-describe('/api/monsters', () => {
-  it('should list monsters', async () => {
-    const res = await request(app).get('/api/monsters');
-    expect(res.statusCode).toEqual(200);
-    expect(res.body.results.length).not.toEqual(0);
-  });
-
-  it('should hit the cache', async () => {
-    redisClient.del('/api/monsters');
-    const clientSet = jest.spyOn(redisClient, 'set');
-    let res = await request(app).get('/api/monsters');
-    res = await request(app).get('/api/monsters');
-    expect(res.statusCode).toEqual(200);
-    expect(res.body.results.length).not.toEqual(0);
-    expect(clientSet).toHaveBeenCalledTimes(1);
-  });
-
-  describe('with name query', () => {
-    it('returns the named object', async () => {
-      const indexRes = await request(app).get('/api/monsters');
-      const name = indexRes.body.results[1].name;
-      const res = await request(app).get(`/api/monsters?name=${name}`);
-      expect(res.statusCode).toEqual(200);
-      expect(res.body.results[0].name).toEqual(name);
-    });
-
-    it('is case insensitive', async () => {
-      const indexRes = await request(app).get('/api/monsters');
-      const name = indexRes.body.results[1].name;
-      const queryName = name.toLowerCase();
-      const res = await request(app).get(`/api/monsters?name=${queryName}`);
-      expect(res.statusCode).toEqual(200);
-      expect(res.body.results[0].name).toEqual(name);
-    });
-  });
-
-  describe('with challenge_rating query', () => {
-    describe('with only one provided challenge rating', () => {
-      it('returns expected objects', async () => {
-        const expectedCR = 0.25;
-        const res = await request(app).get(`/api/monsters?challenge_rating=${expectedCR}`);
-        expect(res.statusCode).toEqual(200);
-
-        const randomIndex = Math.floor(Math.random() * res.body.results.length);
-        const randomResult = res.body.results[randomIndex];
-
-        const indexRes = await request(app).get(`/api/monsters/${randomResult.index}`);
-        expect(indexRes.statusCode).toEqual(200);
-        expect(indexRes.body.challenge_rating).toEqual(expectedCR);
-      });
-    });
-
-    describe('with many provided challenge ratings', () => {
-      it('returns expected objects', async () => {
-        const cr1 = 1;
-        const cr1Res = await request(app).get(`/api/monsters?challenge_rating=${cr1}`);
-        expect(cr1Res.statusCode).toEqual(200);
-
-        const cr20 = 20;
-        const cr20Res = await request(app).get(`/api/monsters?challenge_rating=${cr20}`);
-        expect(cr20Res.statusCode).toEqual(200);
-
-        const bothRes = await request(app).get(`/api/monsters?challenge_rating=${cr1},${cr20}`);
-        expect(bothRes.statusCode).toEqual(200);
-        expect(bothRes.body.count).toEqual(cr1Res.body.count + cr20Res.body.count);
-
-        const randomIndex = Math.floor(Math.random() * bothRes.body.results.length);
-        const randomResult = bothRes.body.results[randomIndex];
-
-        const indexRes = await request(app).get(`/api/monsters/${randomResult.index}`);
-        expect(indexRes.statusCode).toEqual(200);
-        expect(
-          indexRes.body.challenge_rating == cr1 || indexRes.body.challenge_rating == cr20
-        ).toBeTruthy();
-      });
-    });
-  });
-
-  describe('/api/monsters/:index', () => {
-    it('should return one object', async () => {
-      const indexRes = await request(app).get('/api/monsters');
-      const index = indexRes.body.results[0].index;
-      const showRes = await request(app).get(`/api/monsters/${index}`);
-      expect(showRes.statusCode).toEqual(200);
-      expect(showRes.body.index).toEqual(index);
-    });
-    describe('with an invalid index', () => {
-      it('should return one object', async () => {
-        const invalidIndex = 'invalid-index';
-        const showRes = await request(app).get(`/api/monsters/${invalidIndex}`);
-        expect(showRes.statusCode).toEqual(404);
-      });
-    });
-  });
-});
-
 describe('/api/proficiencies', () => {
   it('should list proficiencies', async () => {
     const res = await request(app).get('/api/proficiencies');
